Move StreamdeckContext and Utilities to top level
