+++ Sources/README.md
@@ -1,12 +1,14 @@
 # Directory Structure
 
-- Common: Holds StreamDeck SDK common classes by Elgato.
-- DcsInterface: Provides classes for interacting with DCS (DcsInterface) and maintaining StreamDeck button context (StreamdecContext).
-- Test: Unit tests for C++ code.
-- Vendor: External libraries.
-- Windows: Visual studio project files.
-- com.ctytler.dcs.sdPlugin: Package for the StreamDeck Plugin.
-- MyStreamDeckPlugin - Definition for C++ executable packaged with the StreamDeck Plugin.
+- **Common:** Holds StreamDeck SDK common classes by Elgato.
+- **DcsInterface:** Provides classes for interacting with DCS.
+- **StreamdeckContext:** Provides classes for maintaining StreamDeck button context.
+- **Test:** Unit tests for C++ code.
+- **Utilities:** UDP Socket and string manipulation utilities.
+- **Vendor:** External libraries.
+- **Windows:** Visual studio project files.
+- **com.ctytler.dcs.sdPlugin:** Package for the StreamDeck Plugin.
+- **MyStreamDeckPlugin:** Definition for C++ executable packaged with the StreamDeck Plugin.
 
 ### StreamDeck SDK Reference
 Note this project and directory structure were derived from the SDK sample "CPU": https://github.com/elgatosf/streamdeck-cpu
