+++ Sources/Test/DecimalTest.cpp
@@ -2,7 +2,7 @@
 
 #include "gtest/gtest.h"
 
-#include "../DcsInterface/Utilities/Decimal.cpp"
+#include "../Utilities/Decimal.cpp"
 
 namespace test {
