+++ Sources/Utilities/StringUtilities.cpp
